refactor: Fix linter warnings
