+++ src/constraints/constraint.rs
@@ -1,6 +1,6 @@
 use core::fmt;
 use once_cell::sync::Lazy;
-use std::{collections::HashMap, fmt::Debug, i32};
+use std::{collections::HashMap, fmt::Debug};
 
 use crate::{
     evaluator::{eval_error::EvalError, object::Object},
@@ -310,8 +310,8 @@ impl Constraint for CountConstraint {
             Some(Box::new(IdentitySampler::new(self.min)))
         } else {
             Some(Box::new(UniformSampler::new(
-                self.min as i32,
-                self.max as i32,
+                self.min,
+                self.max,
             )))
         }
     }
