+++ src/constraints/mod.rs
@@ -1,5 +1,5 @@
 pub mod constrainted_type;
-pub mod constraints;
+pub mod constraint;
 pub mod sampler;
 pub mod tests;
 mod util;
