+++ src/constraints/tests.rs
@@ -1,8 +1,7 @@
 #[cfg(test)]
 mod constraint_test {
     use crate::{
-        evaluator::{context::Context, evaluator::evaluate},
-        parser::{parser::Parser, parser_error::ParserError},
+        evaluator::{self, context::Context}, parser::{Parser, parser_error::ParserError}
     };
 
     #[test]
@@ -12,7 +11,7 @@ mod constraint_test {
         match parser.parse() {
             Ok(program) => {
                 let mut context = Context::default();
-                evaluate(program, &mut context).unwrap();
+                evaluator::evaluate(program, &mut context).unwrap();
                 let data_type = context.get_type("even_positive_int");
                 assert!(data_type.is_some());
                 assert_eq!(data_type.unwrap().constraints.len(), 2);
@@ -28,7 +27,7 @@ mod constraint_test {
         match parser.parse() {
             Ok(program) => {
                 let mut context = Context::default();
-                evaluate(program, &mut context).unwrap();
+                evaluator::evaluate(program, &mut context).unwrap();
                 let data_type = context.get_type("even_positive_int");
                 assert!(data_type.is_some());
                 // TODO: add proper assertions
