+++ src/lib.rs
@@ -1,7 +1,7 @@
 pub mod constraints;
 pub mod enumeration;
 pub mod evaluator;
-pub mod generator;
+pub mod generation;
 pub mod lexer;
 pub mod parser;
 pub mod template;
